+++ crates/cli-tools/CHANGELOG.md
@@ -11,7 +11,7 @@
 
 - Add `cmd::status()` to execute a command and return its error code
 - Add `cmd::exit_status()` to execute a command and exit with its error code in case of error
-- Add `action::PlatformUpdate` an optional file for A/B platform update
+- Add an optional file to `action::PlatformUpdate` for A/B platform update
 - Make `action::PlatformUpdate` fields public
 - Add `action::PlatformInfo::print()` to print the platform information
 - Add `--allow=unused-crate-dependencies` to `RUSTFLAGS` when building prod applets to work around a